Move Status > Address & Ports population to `Creating` state processing
Previously the IP and Ports of the GameServer were populated only when
the GameServer is set to Ready.

This information may be needed by the GameServer process before the
process actually wants to mark itself as Ready.

Therefore, port and address population has been moved to the `Creating` state
processing, as that is where the Pod is created, and therefore we have both
address and port information.

`GameServer` events now look like this:

```
Events:
  Type    Reason          Age   From                   Message
  ----    ------          ----  ----                   -------
  Normal  PortAllocation  17s   gameserver-controller  Port allocated
  Normal  Creating        16s   gameserver-controller  Pod simple-udp-fcf44 created
  Normal  Starting        16s   gameserver-controller  Address and Port populated
  Normal  Ready           15s   gameserver-controller  SDK.Ready() executed
```

Closes #293